MOVE CreateSubscription CLASS INTO ITS OWN FILE
- in quizzes/api.py
- in quizzes/helpers/createsubscription.py
